Centralize action integration admission
Amp-Thread-ID: https://ampcode.com/threads/T-01a0055a-3294-703a-b945-abaa212f066d
